fitness/fitplan__/requirements.txt
streamlit>=1.40.0
pyjwt>=2.9.0
python-dotenv>=1.0.1
flask>=2.3.0
flask-cors>=4.0.0
requests>=2.31.0
